slimmed down the block class and moved one offs to helper
